refactor(diffusion_planner): separate core logic (#12151)
* Separated the core logic


* Removed unnecessary includes and using


* Fixed build


* Fixed an error message


* Fixed route_ptr


* Fixed route_ptr_


* Fixed route_ptr_


* Fixed


* Fixed load_model and header


* Fixed to 2026


* Fixed comments in planning/autoware_diffusion_planner/include/autoware/diffusion_planner/diffusion_planner_node.hpp


---------
