+++ planning/autoware_diffusion_planner/src/diffusion_planner_node.cpp
@@ -72,6 +72,11 @@ DiffusionPlanner::DiffusionPlanner(const rclcpp::NodeOptions & options)
   time_keeper_ = std::make_shared<autoware_utils::TimeKeeper>(debug_processing_time_detail_pub_);
 
   set_up_params();
+  vehicle_info_ = autoware::vehicle_info_utils::VehicleInfoUtils(*this).getVehicleInfo();
+
+  // Create core instance
+  core_ = std::make_unique<DiffusionPlannerCore>(params_, vehicle_info_);
+
   turn_indicator_manager_.set_hold_duration(
     rclcpp::Duration::from_seconds(params_.turn_indicator_hold_duration));
   turn_indicator_manager_.set_keep_offset(params_.turn_indicator_keep_offset);
@@ -85,7 +90,6 @@ DiffusionPlanner::DiffusionPlanner(const rclcpp::NodeOptions & options)
     }
   } catch (const std::exception & e) {
     RCLCPP_ERROR_STREAM(get_logger(), e.what() << ". Inference will be disabled.");
-    tensorrt_inference_.reset();
     diagnostics_inference_->update_level_and_message(DiagnosticStatus::ERROR, e.what());
     diagnostics_inference_->publish(get_clock()->now());
     if (params_.build_only) {
@@ -93,8 +97,6 @@ DiffusionPlanner::DiffusionPlanner(const rclcpp::NodeOptions & options)
       std::exit(EXIT_FAILURE);
     }
   }
-
-  vehicle_info_ = autoware::vehicle_info_utils::VehicleInfoUtils(*this).getVehicleInfo();
 
   timer_ = rclcpp::create_timer(
     this, get_clock(), rclcpp::Rate(params_.planning_frequency_hz).period(),
@@ -148,12 +150,9 @@ void DiffusionPlanner::set_up_params()
 
 void DiffusionPlanner::load_model()
 {
-  utils::check_weight_version(params_.args_path);
-  normalization_map_ = utils::load_normalization_stats(params_.args_path);
   diagnostics_inference_->update_level_and_message(DiagnosticStatus::WARN, "Loading model");
   diagnostics_inference_->publish(get_clock()->now());
-  tensorrt_inference_ = std::make_unique<TensorrtInference>(
-    params_.model_path, params_.plugins_path, params_.batch_size);
+  core_->load_model();
   diagnostics_inference_->update_level_and_message(DiagnosticStatus::OK, "Model loaded");
   diagnostics_inference_->publish(get_clock()->now());
 }
@@ -195,12 +194,13 @@ SetParametersResult DiffusionPlanner::on_parameter(
       rclcpp::Duration::from_seconds(params_.turn_indicator_hold_duration));
     turn_indicator_manager_.set_keep_offset(params_.turn_indicator_keep_offset);
 
+    core_->update_params(params_);
+
     if (args_path_changed || model_path_changed || batch_size_changed) {
       try {
         load_model();
       } catch (const std::exception & e) {
         RCLCPP_ERROR_STREAM(get_logger(), e.what() << ". Failed to reload model.");
-        tensorrt_inference_.reset();
         SetParametersResult result;
         result.successful = false;
         result.reason = e.what();
@@ -224,247 +224,10 @@ SetParametersResult DiffusionPlanner::on_parameter(
   return result;
 }
 
-std::optional<FrameContext> DiffusionPlanner::create_frame_context()
-{
-  autoware_utils_debug::ScopedTimeTrack st(__func__, *time_keeper_);
-  auto objects = sub_tracked_objects_.take_data();
-  auto ego_kinematic_state = sub_current_odometry_.take_data();
-  auto ego_acceleration = sub_current_acceleration_.take_data();
-  auto traffic_signals = sub_traffic_signals_.take_data();
-  auto temp_route_ptr = route_subscriber_.take_data();
-  auto turn_indicators_ptr = sub_turn_indicators_.take_data();
-
-  route_ptr_ = (!route_ptr_ || temp_route_ptr) ? temp_route_ptr : route_ptr_;
-
-  TrackedObjects empty_object_list;
-
-  if (params_.ignore_neighbors) {
-    objects = std::make_shared<TrackedObjects>(empty_object_list);
-  }
-
-  if (
-    !objects || !ego_kinematic_state || !ego_acceleration || !route_ptr_ || !turn_indicators_ptr) {
-    RCLCPP_WARN_STREAM_THROTTLE(
-      get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
-      "There is no input data. objects: "
-        << (objects ? "true" : "false")
-        << ", ego_kinematic_state: " << (ego_kinematic_state ? "true" : "false")
-        << ", ego_acceleration: " << (ego_acceleration ? "true" : "false")
-        << ", route: " << (route_ptr_ ? "true" : "false")
-        << ", turn_indicators: " << (turn_indicators_ptr ? "true" : "false"));
-    return std::nullopt;
-  }
-
-  if (traffic_signals.empty()) {
-    RCLCPP_WARN_THROTTLE(
-      this->get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
-      "no traffic signal received. traffic light info will not be updated");
-  }
-
-  Odometry kinematic_state = *ego_kinematic_state;
-  if (params_.shift_x) {
-    kinematic_state.pose.pose =
-      utils::shift_x(kinematic_state.pose.pose, vehicle_info_.wheel_base_m / 2.0);
-  }
-
-  // Get transforms
-  const geometry_msgs::msg::Pose & pose_base_link = kinematic_state.pose.pose;
-  const Eigen::Matrix4d ego_to_map_transform = utils::pose_to_matrix4d(pose_base_link);
-  const Eigen::Matrix4d map_to_ego_transform = utils::inverse(ego_to_map_transform);
-
-  // Update ego history
-  ego_history_.push_back(kinematic_state);
-  if (ego_history_.size() > static_cast<size_t>(EGO_HISTORY_SHAPE[1])) {
-    ego_history_.pop_front();
-  }
-
-  // Update turn indicators history
-  turn_indicators_history_.push_back(*turn_indicators_ptr);
-  if (turn_indicators_history_.size() > static_cast<size_t>(TURN_INDICATORS_SHAPE[1])) {
-    turn_indicators_history_.pop_front();
-  }
-
-  // Update neighbor agent data
-  agent_data_.update_histories(*objects, params_.ignore_unknown_neighbors);
-  const auto processed_neighbor_histories =
-    agent_data_.transformed_and_trimmed_histories(map_to_ego_transform, NEIGHBOR_SHAPE[1]);
-
-  // Update traffic light map
-  const auto & traffic_light_msg_timeout_s = params_.traffic_light_group_msg_timeout_seconds;
-  preprocess::process_traffic_signals(
-    traffic_signals, traffic_light_id_map_, this->now(), traffic_light_msg_timeout_s);
-
-  // Create frame context
-  const rclcpp::Time frame_time(ego_kinematic_state->header.stamp);
-  const FrameContext frame_context{
-    *ego_kinematic_state, *ego_acceleration, ego_to_map_transform, processed_neighbor_histories,
-    frame_time};
-
-  return frame_context;
-}
-
-InputDataMap DiffusionPlanner::create_input_data(const FrameContext & frame_context)
-{
-  autoware_utils_debug::ScopedTimeTrack st(__func__, *time_keeper_);
-  InputDataMap input_data_map;
-
-  // random sample trajectories
-  {
-    for (int64_t b = 0; b < params_.batch_size; b++) {
-      const std::vector<float> sampled_trajectories =
-        preprocess::create_sampled_trajectories(params_.temperature_list[b]);
-      input_data_map["sampled_trajectories"].insert(
-        input_data_map["sampled_trajectories"].end(), sampled_trajectories.begin(),
-        sampled_trajectories.end());
-    }
-  }
-
-  const geometry_msgs::msg::Pose & pose_center =
-    params_.shift_x
-      ? utils::shift_x(
-          frame_context.ego_kinematic_state.pose.pose, vehicle_info_.wheel_base_m / 2.0)
-      : frame_context.ego_kinematic_state.pose.pose;
-  const Eigen::Matrix4d ego_to_map_transform = utils::pose_to_matrix4d(pose_center);
-  const Eigen::Matrix4d map_to_ego_transform = utils::inverse(ego_to_map_transform);
-  const auto & center_x = static_cast<float>(pose_center.position.x);
-  const auto & center_y = static_cast<float>(pose_center.position.y);
-  const auto & center_z = static_cast<float>(pose_center.position.z);
-
-  // Ego history
-  {
-    const std::vector<float> single_ego_agent_past =
-      preprocess::create_ego_agent_past(ego_history_, EGO_HISTORY_SHAPE[1], map_to_ego_transform);
-    input_data_map["ego_agent_past"] =
-      utils::replicate_for_batch(single_ego_agent_past, params_.batch_size);
-  }
-  // Ego state
-  {
-    const auto ego_current_state = preprocess::create_ego_current_state(
-      frame_context.ego_kinematic_state, frame_context.ego_acceleration,
-      static_cast<float>(vehicle_info_.wheel_base_m));
-    input_data_map["ego_current_state"] =
-      utils::replicate_for_batch(ego_current_state, params_.batch_size);
-  }
-  // Agent data on ego reference frame
-  {
-    const auto neighbor_agents_past = flatten_histories_to_vector(
-      frame_context.ego_centric_neighbor_histories, MAX_NUM_NEIGHBORS, INPUT_T + 1);
-    input_data_map["neighbor_agents_past"] =
-      utils::replicate_for_batch(neighbor_agents_past, params_.batch_size);
-  }
-  // Static objects
-  // TODO(Daniel): add static objects
-  {
-    std::vector<int64_t> single_batch_shape(
-      STATIC_OBJECTS_SHAPE.begin() + 1, STATIC_OBJECTS_SHAPE.end());
-    auto static_objects_data = utils::create_float_data(single_batch_shape, 0.0f);
-    input_data_map["static_objects"] =
-      utils::replicate_for_batch(static_objects_data, params_.batch_size);
-  }
-
-  // map data on ego reference frame
-  {
-    const std::vector<int64_t> segment_indices = lane_segment_context_->select_lane_segment_indices(
-      map_to_ego_transform, center_x, center_y, NUM_SEGMENTS_IN_LANE);
-    const auto [lanes, lanes_speed_limit] = lane_segment_context_->create_tensor_data_from_indices(
-      map_to_ego_transform, traffic_light_id_map_, segment_indices, NUM_SEGMENTS_IN_LANE);
-    input_data_map["lanes"] = utils::replicate_for_batch(lanes, params_.batch_size);
-    input_data_map["lanes_speed_limit"] =
-      utils::replicate_for_batch(lanes_speed_limit, params_.batch_size);
-  }
-
-  // route data on ego reference frame
-  {
-    const std::vector<int64_t> segment_indices =
-      lane_segment_context_->select_route_segment_indices(
-        *route_ptr_, center_x, center_y, center_z, NUM_SEGMENTS_IN_ROUTE);
-    const auto [route_lanes, route_lanes_speed_limit] =
-      lane_segment_context_->create_tensor_data_from_indices(
-        map_to_ego_transform, traffic_light_id_map_, segment_indices, NUM_SEGMENTS_IN_ROUTE);
-    input_data_map["route_lanes"] = utils::replicate_for_batch(route_lanes, params_.batch_size);
-    input_data_map["route_lanes_speed_limit"] =
-      utils::replicate_for_batch(route_lanes_speed_limit, params_.batch_size);
-  }
-
-  // polygons
-  {
-    const auto & polygons =
-      lane_segment_context_->create_polygon_tensor(map_to_ego_transform, center_x, center_y);
-    input_data_map["polygons"] = utils::replicate_for_batch(polygons, params_.batch_size);
-  }
-
-  // line strings
-  {
-    const auto & line_strings =
-      lane_segment_context_->create_line_string_tensor(map_to_ego_transform, center_x, center_y);
-    input_data_map["line_strings"] = utils::replicate_for_batch(line_strings, params_.batch_size);
-  }
-
-  // goal pose
-  {
-    const auto & goal_pose = route_ptr_->goal_pose;
-
-    // Convert goal pose to 4x4 transformation matrix
-    const Eigen::Matrix4d goal_pose_map_4x4 = utils::pose_to_matrix4d(goal_pose);
-
-    // Transform to ego frame
-    const Eigen::Matrix4d goal_pose_ego_4x4 = map_to_ego_transform * goal_pose_map_4x4;
-
-    // Extract relative position
-    const float x = goal_pose_ego_4x4(0, 3);
-    const float y = goal_pose_ego_4x4(1, 3);
-
-    // Extract heading as cos/sin from rotation matrix
-    const auto [cos_yaw, sin_yaw] =
-      utils::rotation_matrix_to_cos_sin(goal_pose_ego_4x4.block<3, 3>(0, 0));
-
-    std::vector<float> single_goal_pose = {x, y, cos_yaw, sin_yaw};
-    input_data_map["goal_pose"] = utils::replicate_for_batch(single_goal_pose, params_.batch_size);
-  }
-
-  // ego shape
-  {
-    const float wheel_base = static_cast<float>(vehicle_info_.wheel_base_m);
-    const float vehicle_length = static_cast<float>(
-      vehicle_info_.front_overhang_m + vehicle_info_.wheel_base_m + vehicle_info_.rear_overhang_m);
-    const float vehicle_width = static_cast<float>(
-      vehicle_info_.left_overhang_m + vehicle_info_.wheel_tread_m + vehicle_info_.right_overhang_m);
-    std::vector<float> single_ego_shape = {wheel_base, vehicle_length, vehicle_width};
-    input_data_map["ego_shape"] = utils::replicate_for_batch(single_ego_shape, params_.batch_size);
-  }
-
-  // turn indicators
-  {
-    // copy from back to front, and use the front value for padding if not enough history
-    std::vector<float> single_turn_indicators(INPUT_T + 1, 0.0f);
-    for (int64_t t = 0; t < INPUT_T + 1; ++t) {
-      const int64_t index = std::max(
-        static_cast<int64_t>(turn_indicators_history_.size()) - 1 - t, static_cast<int64_t>(0));
-      single_turn_indicators[INPUT_T - t] = turn_indicators_history_[index].report;
-    }
-    input_data_map["turn_indicators"] =
-      utils::replicate_for_batch(single_turn_indicators, params_.batch_size);
-  }
-
-  return input_data_map;
-}
-
 void DiffusionPlanner::publish_first_traffic_light_on_route(
   const FrameContext & frame_context) const
 {
-  const geometry_msgs::msg::Pose & pose_center =
-    params_.shift_x
-      ? utils::shift_x(
-          frame_context.ego_kinematic_state.pose.pose, vehicle_info_.wheel_base_m / 2.0)
-      : frame_context.ego_kinematic_state.pose.pose;
-
-  const double center_x = pose_center.position.x;
-  const double center_y = pose_center.position.y;
-  const double center_z = pose_center.position.z;
-
-  const auto msg = lane_segment_context_->get_first_traffic_light_on_route(
-    *route_ptr_, center_x, center_y, center_z, traffic_light_id_map_);
-
+  const auto msg = core_->get_first_traffic_light_on_route(frame_context);
   pub_traffic_signal_->publish(msg);
 }
 
@@ -557,7 +320,7 @@ void DiffusionPlanner::on_timer()
   diagnostics_inference_->clear();
 
   const rclcpp::Time current_time(get_clock()->now());
-  if (!tensorrt_inference_) {
+  if (!core_->is_model_loaded()) {
     RCLCPP_WARN_THROTTLE(
       get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
       "Model not loaded. Inference is disabled. Check onnx_model_path and args_path parameters.");
@@ -566,7 +329,7 @@ void DiffusionPlanner::on_timer()
     return;
   }
 
-  if (!lane_segment_context_) {
+  if (!core_->is_map_loaded()) {
     RCLCPP_INFO_THROTTLE(
       get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
       "Waiting for map data...");
@@ -575,52 +338,62 @@ void DiffusionPlanner::on_timer()
     return;
   }
 
-  // Prepare frame context and input data for the model
-  const std::optional<FrameContext> frame_context = create_frame_context();
+  // Take data from subscribers
+  auto objects = sub_tracked_objects_.take_data();
+  auto ego_kinematic_state = sub_current_odometry_.take_data();
+  auto ego_acceleration = sub_current_acceleration_.take_data();
+  auto traffic_signals = sub_traffic_signals_.take_data();
+  auto temp_route_ptr = route_subscriber_.take_data();
+  auto turn_indicators_ptr = sub_turn_indicators_.take_data();
+
+  // Prepare frame context using core
+  const std::optional<FrameContext> frame_context = core_->create_frame_context(
+    ego_kinematic_state, ego_acceleration, objects, traffic_signals, turn_indicators_ptr,
+    temp_route_ptr, this->now());
+
   if (!frame_context) {
-    RCLCPP_WARN_THROTTLE(
+    // Log detailed information about missing inputs
+    RCLCPP_WARN_STREAM_THROTTLE(
       get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
-      "No input data available for inference");
+      "There is no input data. objects: "
+        << (objects ? "true" : "false")
+        << ", ego_kinematic_state: " << (ego_kinematic_state ? "true" : "false")
+        << ", ego_acceleration: " << (ego_acceleration ? "true" : "false")
+        << ", route: " << (core_->get_route() ? "true" : "false")
+        << ", turn_indicators: " << (turn_indicators_ptr ? "true" : "false"));
     diagnostics_inference_->update_level_and_message(
       DiagnosticStatus::WARN, "No input data available for inference");
     diagnostics_inference_->publish(current_time);
     return;
   }
 
+  if (traffic_signals.empty()) {
+    RCLCPP_WARN_THROTTLE(
+      this->get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
+      "no traffic signal received. traffic light info will not be updated");
+  }
+
   const rclcpp::Time frame_time(frame_context->frame_time);
-  InputDataMap input_data_map = create_input_data(*frame_context);
+  InputDataMap input_data_map = core_->create_input_data(*frame_context);
 
   publish_debug_markers(input_data_map, frame_context->ego_to_map_transform, frame_time);
 
   publish_first_traffic_light_on_route(*frame_context);
 
-  // Calculate and record metrics for diagnostics using the proper logic
-  const int64_t batch_idx = 0;
-  const int64_t valid_lane_count = postprocess::count_valid_elements(
-    input_data_map["lanes"], LANES_SHAPE[1], LANES_SHAPE[2], LANES_SHAPE[3], batch_idx);
-  diagnostics_inference_->add_key_value("valid_lane_count", valid_lane_count);
-
-  const int64_t valid_route_count = postprocess::count_valid_elements(
-    input_data_map["route_lanes"], ROUTE_LANES_SHAPE[1], ROUTE_LANES_SHAPE[2], ROUTE_LANES_SHAPE[3],
-    batch_idx);
-  diagnostics_inference_->add_key_value("valid_route_count", valid_route_count);
-
-  const int64_t valid_polygon_count = postprocess::count_valid_elements(
-    input_data_map["polygons"], POLYGONS_SHAPE[1], POLYGONS_SHAPE[2], POLYGONS_SHAPE[3], batch_idx);
-  diagnostics_inference_->add_key_value("valid_polygon_count", valid_polygon_count);
-
-  const int64_t valid_line_string_count = postprocess::count_valid_elements(
-    input_data_map["line_strings"], LINE_STRINGS_SHAPE[1], LINE_STRINGS_SHAPE[2],
-    LINE_STRINGS_SHAPE[3], batch_idx);
-  diagnostics_inference_->add_key_value("valid_line_string_count", valid_line_string_count);
-
-  const int64_t valid_neighbor_count = postprocess::count_valid_elements(
-    input_data_map["neighbor_agents_past"], NEIGHBOR_SHAPE[1], NEIGHBOR_SHAPE[2], NEIGHBOR_SHAPE[3],
-    batch_idx);
-  diagnostics_inference_->add_key_value("valid_neighbor_count", valid_neighbor_count);
+  // Calculate and record metrics for diagnostics using core
+  diagnostics_inference_->add_key_value(
+    "valid_lane_count", core_->count_valid_elements(input_data_map, "lanes"));
+  diagnostics_inference_->add_key_value(
+    "valid_route_count", core_->count_valid_elements(input_data_map, "route_lanes"));
+  diagnostics_inference_->add_key_value(
+    "valid_polygon_count", core_->count_valid_elements(input_data_map, "polygons"));
+  diagnostics_inference_->add_key_value(
+    "valid_line_string_count", core_->count_valid_elements(input_data_map, "line_strings"));
+  diagnostics_inference_->add_key_value(
+    "valid_neighbor_count", core_->count_valid_elements(input_data_map, "neighbor_agents_past"));
 
   // normalization of data
-  preprocess::normalize_input_data(input_data_map, normalization_map_);
+  preprocess::normalize_input_data(input_data_map, core_->get_normalization_map());
   if (!utils::check_input_map(input_data_map)) {
     RCLCPP_WARN_THROTTLE(
       get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
@@ -631,8 +404,8 @@ void DiffusionPlanner::on_timer()
     return;
   }
 
-  // Run inference
-  const auto inference_result = tensorrt_inference_->infer(input_data_map);
+  // Run inference using core
+  const auto inference_result = core_->run_inference(input_data_map);
   if (!inference_result.outputs) {
     RCLCPP_WARN_STREAM_THROTTLE(
       get_logger(), *this->get_clock(), constants::LOG_THROTTLE_INTERVAL_MS,
@@ -646,10 +419,8 @@ void DiffusionPlanner::on_timer()
 
   publish_predictions(predictions, *frame_context, frame_time);
 
-  // Publish turn indicators
-  const int64_t prev_report = turn_indicators_history_.empty()
-                                ? TurnIndicatorsReport::DISABLE
-                                : turn_indicators_history_.back().report;
+  // Get previous turn indicator report from core's history
+  const int64_t prev_report = core_->get_previous_turn_indicator_report();
   const auto turn_indicator_command =
     turn_indicator_manager_.evaluate(turn_indicator_logit, frame_time, prev_report);
   pub_turn_indicators_->publish(turn_indicator_command);
@@ -662,7 +433,7 @@ void DiffusionPlanner::on_map(const HADMapBin::ConstSharedPtr map_msg)
 {
   std::shared_ptr<const lanelet::LaneletMap> lanelet_map_ptr =
     autoware::experimental::lanelet2_utils::from_autoware_map_msgs(*map_msg);
-  lane_segment_context_ = std::make_unique<preprocess::LaneSegmentContext>(lanelet_map_ptr);
+  core_->set_map(lanelet_map_ptr);
 }
 
 }  // namespace autoware::diffusion_planner
